+++ src/server/src/core/systemMonitor.ts
@@ -19,6 +19,7 @@ interface CpuInfo {
     arch: string
     usagePercent: number
     coresUsage: number[]
+    usagePercentData: number[]
 }
 
 // 系统信息数据
@@ -36,6 +37,7 @@ interface SystemInfo {
         free: number
         usage: number
         usagePercent: number
+        usagePercentData: number[]
     }
 }
 
@@ -57,9 +59,34 @@ interface MonitorData {
 
 class SystemMonitor {
     private cpuMonitor: InstanceType<typeof CPUMonitor>
+    private cpuUsagePercentData: number[] = []
+    private ramUsagePercentData: number[] = []
+
     constructor() {
         // 创建 CPU 监控实例，可传入自定义采样间隔
         this.cpuMonitor = new CPUMonitor(1000)
+    }
+
+    // 更新 CPU 使用率数据并返回最新数据组
+    private updateCpuUsageData(): number[] {
+        const cpuUsageAvg = this.cpuMonitor.getUsage().avg
+        this.cpuUsagePercentData.push(cpuUsageAvg)
+        // 记录最近数据条目数
+        if (this.cpuUsagePercentData.length > 60) {
+            this.cpuUsagePercentData.shift()
+        }
+        return this.cpuUsagePercentData;
+    }
+
+    // 更新 RAM 使用率数据并返回最新数据组
+    private updateRamUsageData(): number[] {
+        const ramUsage = (1 - os.freemem() / os.totalmem()) * 100
+        this.ramUsagePercentData.push(ramUsage)
+        // 记录最近数据条目数
+        if (this.ramUsagePercentData.length > 60) {
+            this.ramUsagePercentData.shift()
+        }
+        return this.ramUsagePercentData;
     }
 
     /**
@@ -97,12 +124,14 @@ class SystemMonitor {
                 arch: os.arch(),
                 usagePercent: cpuUsage.avg,
                 coresUsage: cpuUsage.cores,
+                usagePercentData: this.updateCpuUsageData(),
             },
             ram: {
                 total: os.totalmem(),
                 free: os.freemem(),
                 usage: os.totalmem() - os.freemem(),
                 usagePercent: (1 - os.freemem() / os.totalmem()) * 100,
+                usagePercentData: this.updateRamUsageData(),
             },
         }