+++ src/web/src/views/TestView.vue
@@ -10,8 +10,7 @@
     </MainTitle>
 
     <!-- 以下是测试内容 -->
-
-
+     
 </template>
 
 <script setup>